Move async fetch logic up out of Map component
Was needed to ensure Map responds instantly to switch of datasets, and
was causing issues in the function returned by createGriddedDataLoader
where griddedTextureInfo could have different value before and after the
await.

This also changes the external api of Map to require a Float32Array of
data instead merely the path to where that could be found. So, it is now
up to App.svelte to handle the async fetch logic (implemented crudely
for now).