Switch to pre-multiplied alpha for Safari/iOS
A bug in Safari/iOS prevents use of premultiplied alpha, see
https://stackoverflow.com/questions/57150436/webgl-does-not-have-an-alpha-background-channel-on-ios-safari

Makes particle color less white and more the color of the background,
probably could be tweaked.

This fix causes alpha issues on Firefox on Linux:
https://stackoverflow.com/questions/46106325/firefox-webgl-2-0-rgba-texture-strange-behavior

The solution given requires disabling premultiplied alpha, which
obviously won't work in this case (we just enabled it to appease
Safari).

Considering officially dropping support for Firefox on Linux and giving
users an alert box with the suggestion to view the page on Chromium,
since there is already a pretty big FPS difference between the two. Not
ideal at all, however.