Resize canvas based on body instead of canvas
Should fix the bug where page loading sometimes causes causes a layout
shift (a race-condition in mounting the menus it seems like?), and so
the canvas gets resized to much a lower resolution that it should be.

Also fixes the issue when resizing with a menu open would cause the
canvas to again be a lower resolution than it should when that menu was
closed.