Enable webgl2 if available to support more devices
Some phones support webgl2 but not the promoted extensions in webgl that
are needed to render the map. Switching to webgl2 if available expands
our support to those devices.

List of relevant promoted extensions:
- OES_texture_float
- OES_texture_half_float_linear
- OES_texture_half_float

The last two will be used in a not-yet-implemented workaround for
devices that don't support OES_texture_float_linear.