Don't start with Float32Arrays if we don't need to
Partially reverts 8cede0ac to avoid potential crashes on devices that
don't support Float32 interpolation. (Not 100% sure if crash actually
happens since I do not have a test device for this, but this makes the
mental model simpler anyhow: Float16 is default, Float32 is used only if
explicitly requested by web component user or in the to-be-detected iOS
<= 12 case for the website.) Uses a better default value this time
around.