Switch date from a bound variable to a store
Workaround for the behavior in the second REPL here:

https://github.com/sveltejs/svelte/issues/4265#issuecomment-731783590

causing two gridded data fetches to happen per change instead of one.
(And fast enough that the Fetcher object doesn't seem to be able to
abort them? Might be a separate issue.)

Also a good time to learn how to use stores, I guess. I think having the
bound variables is clearer in this case so we may want to revert this
once the Svelte bug is fixed, but stores are still pretty neat.