Add datasets functionality to Variables menu
Adds the functionality of the Gridded Datasets menu into the Variables
menu, so that choosing variables and timepoints is now handled through
this menu.

Removes the Gridded Datasets menu.

The time-step buttons are included here, but if the time-slider menu is
brought online in time for the demo, they should be moved to that menu
instead.

Implementation note: I recognize that the current implementation of this
menu is verbose and contains redundancies. However, I wanted to get this
menu functional and looking polished in time for the demo. An imminent
to-do once the demo is done is to refactor both the Variables menu and
the Projections menu to use styled radio buttons and Svelte "each"
blocks, in order to simplify the code, make these menus more
keyboard-accessible, and facilitate future user-defined variables.