Implement loading button and slider for time range
Mostly working, still need some tweaks to hide/show the load button and
slider when appropriate and figure the ergonomics of enabling/disabling
particles when loading (since they need a lot of data to download but
don't really help the sliding-through-time-experience).

Also need to figure out how to disable the range slider binding when
changing date to value outside of the pre-loaded range.