Fix jank in mouse panning of globe
Jank was introduced by 9775886f because binding the centerLongitude and
centerLatitude variables to the RangeSlider causes them to clamp to the
nearest step on every update.

Fix is to avoid updating the `value` prop in Rangeslider if the current
`value`'s closest step is already in `values`. Somewhat sketch, yes.