Majorly refactor inventory (again) for simplicity
Now, the dataset variables are objects instead of paths, and the
inventory is an array instead of a object mapping paths to dataset info.

The reduces the need to query the inventory for information in various
components if they already have the dataset variable (e.g.
griddedDataset replacing griddedDatasetInfo in many cases).

Furthermore, more constants relating to particle datasets have been moved
to the inventory, and particle datasets are again separated from gridded
ones (partial revert of 37c655b). Logic for fetching the two files
needed for particle datasets has been moved into the Fetcher class.

Finally, the contents of inventory.json now directly dictate the content
of the Variables menu instead of needing an additional manual mapping of
dataset name to path within the component.