Keep legend info in sync when awaiting download
Before, info (like the title of the legend and the colormap displayed)
would be mixed when switching datasets as griddedDataset would update
before griddedColormap and griddedDomain. Also pre-emptively account for
this in the StreamlinesLegend.

Downside to this is that everything in the legend components essentially
runs twice... (too fast to notice hopefully)

Calling the createAxis function reactively removes the need for onMount,
as it gets triggered when svgScale changes from undefined to the
element.