Modify GFS script to use forecasts + analysis
Allows us to fill the gaps between the 4-times-daily GFS forecasts to
offer 1 hour resolution instead of 6.

Will want to add some indication to frontend as to which data is
forecasted and when that forecast was created by GFS. This commit also
adds a few new dataset props to inventory.json to help with this.