Fix original FEVer colormaps
Allows us to copy the colormaps exactly as they are defined in the
original FEVer (and earth.nullschool.net) instead of trying to
approximate them.

Credit to Julien P. Nicolas (and other FEVer 1 collaborators) for all
modifications to the code and colormaps in earth.js; I only removed the
alpha argument in the function returned from segmentedColorScales.

The `0.66` constant comes from the OVERLAY_ALPHA constant in the
original earth.nullschool.net codebase. This made the displayed colors
darker than the what was depicted in the legend in the old applications;
this problem is resolved in the implementation here.