Hotfix prevent crash when switching datasets
Only occurs if switching from dataset with 0 as limit of domain to a
dataset with temperature units and a different valid date (i.e. sea
surface temperature), and only consistently crashes in dev: sometimes
but not always crashes in prod...

Very strange. Maybe something to do with the `await tick()`s in
App.svelte.