Switch to stricter implementation of throttle
Had issues where (maybe? maybe not?) over 119 calls were performed in a
minute according to our journalctl logs, even with the `strict` option
in p-throttle. Let's try our luck with this library? From the same
author of the strict implementation in p-throttle but with update to
improve accuracy using nanotimer.

Found in the following comment:

https://github.com/sindresorhus/p-throttle/pull/27#issuecomment-779255087