Fix crash when menu is opened during page load
If the dataset specified in the hash has incompatible units with the
default initial dataset, opening the menu during page load can cause
`convert` to be called with mismatched dataset and unit, causing the Qty
library to throw an exception and crash the whole program.

This fixes that, kinda hackily.