Fix text not being vertically centered on Windows
In effect switches the font on used on Windows from Segoe UI to Arial.

Temporary solution to issue caused by uneven ascenders/descenders in
Segoe UI. We can either decide on a single font to use across OSes or
wait for https://www.w3.org/TR/css-inline-3/#leading-trim to be
implemented in browsers.