Fix dataset switch causing permanent date change
For instance, consider the following scenario:

1. User loads initial page with the 2m temperature dataset (1 hour interval)
2. User switches to sea surface temperature dataset (24 hour interval),
causing the date to change to the nearest valid date
3. User switches back to the 2m temperature dataset

The expectation is that after switching back to the initial dataset, the
date is the same as it was initially. However, the old behavior was that
the date would remain as it was after it was modified to be a valid date
for sea surface temperature.

The old values of date in the above scenario were:

1. 2021-05-21T16:00:00.000Z
2. 2021-05-20T00:00:00.000Z
3. 2021-05-20T00:00:00.000Z

The new values:

1. 2021-05-21T16:00:00.000Z
2. 2021-05-20T00:00:00.000Z
3. 2021-05-21T16:00:00.000Z