Fix NetCDF conversion script for Windows
A bug in `ncdump` causes backslashes to be removed from the path given
as a command line argument. This wasn't a problem when calling the
conversion script manually as long as paths were typed with forward
slashes, but causes issues when the conversion script is called from
another node script because `path.join` uses backslashes as the default
path separator on Windows. Specifically, this would cause the output
files to be zero bytes, as no data would be piped.

This will be fixed in the next version of netCDF:

https://github.com/Unidata/netcdf-c/pull/1989