Keep zoom proportional to screen dimension
Adds an option to Map component to toggle between landscape and
portrait-based zoom (zoom stays consistent when width or height changes,
respectively) and assume that small screen width ("mobile") means that
device is in portrait mode. Also exposes a new prop synced to the aspect
ratio of the canvas for use in Controls.