Implement rudimentary player for timelapses
Largely based on Mediabar code removed in 0a37212a0

Currently tries to load and loop through the entire dataset from
earliest date to latest, which is fine for local testing (where only a
few times are available) but obviously not for prod. Will need to add
proper settings that determine and limit the default date ranges.