Double the supported width for griddedData in Map
Displays every other point to stay within the limits of the max texture
size for the specific WebGL implementation. Doesn't affect the values
read by JavaScript (like the labels for markers).

Causes a ~1 second freeze of the UI when loading/splitting too-large
gridded data for the first time, but caches the split result so that
subsequent loads are fast. We may want to try moving the splitting task
to a web worker?

In practice, this allows the sea surface temperature dataset to be
viewable on Chrome on Android, which is a significant percentage of all
browsers used worldwide. Currently, none of our other datasets have
dimensions large enough to require this workaround.

Removes the check in DeveloperOnlyTools as it no longer has any effect.