Make it feasible to serve app from prod server
Patches Vite so that instead of copying all the files from `public/data` to
`dist/data`, it symlink the directories instead.

Before, the large size of `public/data` made it impractical to build the
site anywhere where `public/data` was not empty, such as the prod server
and local development environments, as copying all the files would take
a substantial amount of time. This wasn't a problem previously because
`public/data` would always be empty when built on Netlify.

Now, `npm run build:app` works well on all environments.

Note: to make this patch as small as possible so that is easy to carry
forward to new Vite versions, a side-effect is that *any* directory
named `data` that has `public` as an ancestor will be symlinked, not
just `public/data`. This doesn't affect anything now, but may come back
to bite us later.