Simplify use of webkit-prefixed CSS rules
Removes the tap highlight color in all cases except for links; we
generally have an equivalent visual effect already implemented where
the role of the highlight is important.

Fixes forgetting to remove the highlight in some places, like the new
zoom buttons, and restores the tap highlight to links that were
previously inheriting the transparent value from a parent.