Generate compressed `.fp16` files
By default, uses gzip to avoid issues when using Firefox when developing
on localhost, but this can be overriden with the `FEV2R_USE_BROTLI`
environment variable (e.g. for Netlify). Using brotli is hardcoded to be
the default on the backend production server.

Reduces total data file size by about half and reduces work for the
webserver.

Temporarily has the frontend do a check for the existance of an
uncompressed `.fp16` file so that we can convert the existing files on
the production server without downtime.