Prevent silent failures by long-running units
Sets a timeout so that unit is always inactive by the next trigger (with
ten minutes to spare). Before, units that were still running would cause
trigger to not do anything.

More robustly solves the issue addressed by parent commit (which is
still useful as `got` will retry after the download timeout).

Could switch to RuntimeMaxSec and remove Type=oneshot when we update to
a newer version of systemd.