Generate Brotli-compressed `.fp16` files
Reduces stored data size by about half and avoids the web server needing
to compress on the fly. As a trade-off, conversion scripts are many times
slower than before as Brotli defaults to the max compression level.

Requires using HTTPS for localhost on Firefox, waiting on
https://bugzilla.mozilla.org/show_bug.cgi?id=1222541

See comment in vite.config.js on how to generate local cert for HTTPS