Rewrite Calendar using new Date wrapper class
Moves much of the shared logic between the pickers into Calendar so that
the pickers are now mostly declarative and no longer have UTC-specific
variants. The intention being to reduce duplication and make it easier
to switch to the Temporal API when that comes out.

Also includes some changes to the validDate function, and removes the
tooltips for boxes.