Have all modules use the same retry/timeout values
Apparently, only in Workers is the the dynamic import getting new values
when we modify files; otherwise those are just cached (per ES spec?), so
having them be imported like they were here isn't really helpful. Would
need a dedicated function to make a worker to get the new values, so
just avoid that for now and give every rabbit the same value for retry
and timeout.