Reduce compression level to 6
Over twice as fast, which allows us the possibility of more datasets in
the future without worrying about overloading the CPU of our shared
production server. Compression rates aren't too bad (~2% larger than
level 9).

Chosen as a sweet spot based on benchmarking various compression levels
with the data created by an initial `npm run data` (which we are
replacing with this rewrite!).