Use range requests for GEOS files
Avoids us having to download the whole file when we only use a small
part of it. Requires netCDF-C v4.8.2 (not yet released) because of a bug
in v4.8.1 (https://github.com/Unidata/netcdf-c/issues/2122).

Doesn't work for OSCAR netCDF files ("Unknown file format", possibly due
to their gzip compression?) but not a big deal since we use most of
those files.