Use multiple data textures for higher resolutions
If the max texture size for the webgl instance is larger than each
dimension of the gridded data, render as normal. Otherwise, the gridded
data is re-gridded into one or more textures with width equal to the max
texture size and the `texelFetch` function is used. Data rendered this
way cannot be linearly filtered to appear smooth.

Removes the hack used to render sea surface temperature data on Android
Chrome and the MAX_TEXTURE_SIZE variable previously exported from the
Map component/webcomponent.