Fix GFS data after FTPPRD discontinuation
In theory, this shouldn't have affected us because we should auto-switch
to NOMADS if FTPPRD is not available. However, the logic in 80e3932df
was bad and we somehow ended up in a state where it was trying to switch
to FTPPRD instead of away from it. So, let's just tear out the switching
logic entirely since we no longer have two URLs that work.

See https://www.weather.gov/media/notification/pdf_2025/scn25-35_discontinue_https_access_for_FTPPRD.pdf